+++ ats-mini/Storage.cpp
@@ -65,8 +65,20 @@ void eepromInvalidate()
   // use to free up memory and avoid memory leaks
   EEPROM.begin(EEPROM_SIZE);
   EEPROM.write(EEPROM_BASE_ADDR, 0x00);
+  EEPROM.write(EEPROM_VER_ADDR + 2, 0x01);
   EEPROM.commit();
   EEPROM.end();
+}
+
+// Return true first time after the settings have been reset
+bool eepromFirstRun()
+{
+  EEPROM.begin(EEPROM_SIZE);
+  bool firstRun = EEPROM.read(EEPROM_VER_ADDR + 2);
+  if(firstRun) EEPROM.write(EEPROM_VER_ADDR + 2, 0x00);
+  EEPROM.end();
+
+  return(firstRun);
 }
 
 // Check EEPROM contents against EEPROM_VERSION
@@ -153,6 +165,7 @@ void eepromSaveConfig()
   EEPROM.write(addr++, (uint8_t)zoomMenu);       // Stores the current Zoom Menu setting
   EEPROM.write(addr++, scrollDirection<0? 1:0);  // Stores the current Scroll setting
   EEPROM.write(addr++, utcOffsetIdx);            // Stores the current UTC Offset
+  EEPROM.write(addr++, currentSquelch);          // Stores the current Squelch value
   EEPROM.commit();
 
   addr = EEPROM_SETP_ADDR;
@@ -226,6 +239,7 @@ void eepromLoadConfig()
   zoomMenu       = (bool)EEPROM.read(addr++);    // Reads stored Zoom Menu setting
   scrollDirection = EEPROM.read(addr++)? -1:1;   // Reads stored Scroll setting
   utcOffsetIdx   = EEPROM.read(addr++);          // Reads the current UTC Offset
+  currentSquelch = EEPROM.read(addr++);          // Reads the current Squelch value
 
   addr = EEPROM_SETP_ADDR;
   for(int i=0 ; i<getTotalBands() ; i++)